+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Added
-
+- Added related works project overview page [#700]
 
 ## Updated